requirements.txt
fastapi
pydantic
uvicorn
kafka-python
psycopg2